# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sede",
        "description": "Listado de las URL de la Api Sede",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "remember": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso, redirige según rol"
                    },
                    "422": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar nuevo usuario",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "rol",
                                    "nombre",
                                    "apellido"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "nuevo@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "rol": {
                                        "type": "string",
                                        "example": "administrador"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro exitoso y redirección"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
...
        "/dashboard/admin": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard del administrador",
                "operationId": "1b19b7906ce3cd3a3b914b44e0a4d52f",
                "responses": {
                    "200": {
                        "description": "Vista del dashboard del administrador"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}